refactor(frontend): split MessageDetails section components
Extract RequestHeadersSection, ModelParamsSection, and the formatParamValue
helper into MessageDetailsSections.tsx so MessageDetails.tsx stays under the
300-line file budget. No behavior change — same tests, same render output.
